+++ web/src/locales/en.ts
@@ -1291,6 +1291,7 @@ This delimiter is used to split the input text into several text pieces echo of 
       agent: 'Agent',
       agentDescription:
         'Builds agent components equipped with reasoning, tool usage, and multi-agent collaboration. ',
+      maxRecords: 'Max records',
     },
     llmTools: {
       bad_calculator: {
