+++ web/src/locales/zh.ts
@@ -1244,6 +1244,7 @@ General：实体和关系提取提示来自 GitHub - microsoft/graphrag：基于
       query: '查询变量',
       agent: 'Agent',
       agentDescription: '构建具备推理、工具调用和多智能体协同的智能体组件。',
+      maxRecords: '最大记录数',
     },
     footer: {
       profile: 'All rights reserved @ React',
