+++ web/src/pages/agent/form-sheet/next.tsx
@@ -19,8 +19,8 @@ import { useHandleNodeNameChange } from '../hooks/use-change-node-name';
 import OperatorIcon from '../operator-icon';
 import useGraphStore from '../store';
 import { needsSingleStepDebugging } from '../utils';
+import { FormConfigMap } from './form-config-map';
 import SingleDebugDrawer from './single-debug-drawer';
-import { useFormConfigMap } from './use-form-config-map';
 
 interface IProps {
   node?: RAGFlowNodeType;
@@ -43,8 +43,6 @@ const FormSheet = ({
 }: IModalProps<any> & IProps) => {
   const operatorName: Operator = node?.data.label as Operator;
   const clickedToolId = useGraphStore((state) => state.clickedToolId);
-
-  const FormConfigMap = useFormConfigMap();
 
   const currentFormMap = FormConfigMap[operatorName];
