+++ web/src/pages/agent/form/exesql-form/use-submit-form.ts
@@ -3,15 +3,14 @@ import { useCallback } from 'react';
 import { z } from 'zod';
 
 export const ExeSQLFormSchema = {
-  llm_id: z.string().min(1),
+  sql: z.string(),
   db_type: z.string().min(1),
   database: z.string().min(1),
   username: z.string().min(1),
   host: z.string().min(1),
   port: z.number(),
   password: z.string().min(1),
-  loop: z.number(),
-  top_n: z.number(),
+  max_records: z.number(),
 };
 
 export const FormSchema = z.object({
